#12 feat : 공통으로 사용할 수 있는 함수 분리
- 별 찍는 함수
- 문자열 입력 받는 함수
- 숫자만 입력 받는 함수 (예외 처리)
